requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
pandas>=1.4.0
pillow>=9.0.0
ipykernel>=6.0.0
numpy>=1.21.0
seaborn>=0.11.0
tqdm>=4.65.0
fpdf2>=2.7.0
gradio>=4.0.0